Extract AccountDropdown from header
This commit reimplementing ProjectDropdown as Dropdown component.
